pyproject.toml
[project]
name = "backend-bugdify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4,<5.3",
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.1,<4.0",
    "python-dotenv>=1.1.1",
]
